app/requirements.txt
Django==4.0.3
strawberry-graphql==0.102.2
strawberry-graphql-django==0.3rc1
strawberry-django-jwt==0.2.0
typing-extensions==4.1.1
